opening-book-service/requirements.txt
fastapi==0.112.2
uvicorn==0.30.5
httpx==0.27.0
requests
pydantic==2.8.2
python-chess==1.999
psycopg2-binary==2.9.9
asyncpg==0.29.0
redis
opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-fastapi
opentelemetry-instrumentation-httpx
opentelemetry-instrumentation-requests
opentelemetry-instrumentation-asyncpg
opentelemetry-instrumentation-redis
opentelemetry-instrumentation-system-metrics
